requirements.txt
Flask==3.0.3
pandas==2.0.3
pygal==3.0.4
python-dotenv==1.0.1
Requests==2.32.3
